backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
PyMySQL==1.1.1
python-multipart==0.0.17
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
pymysql==1.1.1
httpx==0.27.2